requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
requests
tqdm
python-dotenv
